Mirror upstream elastic/elasticsearch#134799 as single snapshot commit for AI review BASE=5ce0c649d6b0ebe7c731eb23157489e8d3c46633 HEAD=8c7939c511b854052233c03dc866f469270829d3 Branch=main
